requirements.txt
streamlit
openai
pandas
python-dotenv
fpdf
portkey-ai>=1.14.0